auth/db/requirements.txt
alembic~=1.13.3
pydantic~=2.9.2
pydantic-settings~=2.6.1
python-dotenv~=1.0.1
psycopg2-binary==2.9.10
#sqlalchemy[asyncio]==2.0.36
werkzeug==3.1.2
asyncpg==0.30.0